backend/pyproject.toml
# backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Backend API service for Society Bug Bounty platform, built with FastAPI, featuring OAuth authentication, WebSocket support, and Celery task processing."
requires-python = ">=3.13"
dependencies = [
    "alembic==1.17.2",
    "asyncpg==0.30.0",
    "fastapi==0.121.2",
    "starlette==0.49.0",
    "pydantic[email]==2.8.2",
    "pydantic-settings==2.12.0",
    "sqlalchemy==2.0.44",
    "uvicorn==0.37.0",
    "httpx==0.28.1",
    "httpx-oauth==0.16.1",
    "aiofiles==25.1.0",
    "typer==0.20.0",
    "fastapi-users[sqlalchemy]==15.0.1",
    "fastapi-users-db-sqlalchemy==7.0.0",
    "python-slugify==8.0.4",
    "python-magic==0.4.27",
    "Pillow==12.0.0",
    "python-socketio==5.14.3",
    "python-jose[cryptography]==3.3.0",
    "psycopg==3.2.1",
    "celery[redis]==5.5.3",
    "flower==2.0.1",
    "redis>=5.0.0",
    "gunicorn==23.0.0",
]

[project.scripts]
manage-organizations = "app.cli.manage_organizations:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
